Organize server sync code
Separated to:
* Sync: initiated by DB transaction, wrote to sync
* SyncWatch: initiated by sync, trigger extensions
* StateWatch: initiated by sync, trigger state/monitoring updates
